refactor(desktop): split session refresh and snapshot helpers
