+++ apps/desktop/src/platform/loopal-backend/node/runtime/loopal-live-session.ts
@@ -13,6 +13,7 @@ import { LoopalLiveAttention } from '../attention/loopal-live-attention'
 import { loadSessionDetail } from '../sessions/loopal-session-snapshot'
 import { LoopalMetaHubWatcher } from '../federation/loopal-metahub-watcher'
 import { type SessionRuntimeHandle } from './session-runtime-registry'
+import { LoopalSessionRefresh } from './loopal-session-refresh'
 interface LiveSessionSink {
   event(event: DesktopEvent): void
   summary(summary: SessionSummary): void
@@ -21,11 +22,8 @@ interface LiveSessionSink {
 export class LoopalLiveSession {
   private readonly projector: LoopalEventProjector
   private readonly attention: LoopalLiveAttention
+  private readonly refreshes: LoopalSessionRefresh
   private detailValue?: SessionDetail
-  private refreshing: Promise<void> | undefined
-  private refreshPending = false
-  private refreshShouldEmit = false
-  private refreshTimer: ReturnType<typeof setTimeout> | undefined
   private readonly metaHub: LoopalMetaHubWatcher
   private active = true
 
@@ -36,16 +34,17 @@ export class LoopalLiveSession {
     private readonly sink: LiveSessionSink,
   ) {
     this.attention = new LoopalLiveAttention(runtime, now, sink.event)
+    this.refreshes = new LoopalSessionRefresh((emit) => this.refreshSnapshot(emit))
     this.projector = new LoopalEventProjector(now, {
       append: (entry) => this.append(entry),
       appendAgent: (entry, agentId) => this.appendToAgent(agentId, entry),
       updateSession: (status, attention) => this.update(status, attention),
-      overflow: () => this.requestAuthoritativeRefresh(),
+      overflow: () => this.refreshes.requestImmediately(),
       attention: (kind, value, agentId) => this.attention.accept(kind, value, agentId),
       artifacts: (paths, agentId) => this.recordArtifacts(paths, agentId),
     })
     this.metaHub = new LoopalMetaHubWatcher(
-      runtime.host, now, () => this.detailValue?.metaHub, () => this.refresh(true),
+      runtime.host, now, () => this.detailValue?.metaHub, () => this.refreshes.request(true),
     )
   }
 
@@ -62,22 +61,22 @@ export class LoopalLiveSession {
   }
 
   async initialize(): Promise<void> {
-    await this.refresh(false)
+    await this.refreshes.request(false)
     this.metaHub.start()
   }
 
   resync(): Promise<void> {
-    return this.refresh(true)
+    return this.refreshes.request(true)
   }
 
   accept(method: string, params: unknown): void {
     if (!this.active) return
     if (method === 'agent/event') {
       this.projector.accept(params)
-      this.invalidate()
+      this.refreshes.invalidate()
     }
     else if (method === 'view/resync_required') {
-      void this.refresh(true).catch(() => undefined)
+      void this.refreshes.request(true).catch(() => undefined)
     }
   }
 
@@ -103,39 +102,15 @@ export class LoopalLiveSession {
 
   dispose(): void {
     this.active = false
-    if (this.refreshTimer) clearTimeout(this.refreshTimer)
-    this.refreshTimer = undefined
+    this.refreshes.dispose()
     this.metaHub.dispose()
   }
 
   retire(): readonly DesktopEvent[] {
     this.active = false
-    if (this.refreshTimer) clearTimeout(this.refreshTimer)
-    this.refreshTimer = undefined
+    this.refreshes.dispose()
     this.metaHub.dispose()
     return this.attention.retire()
-  }
-
-  private refresh(emit: boolean): Promise<void> {
-    this.refreshPending = true
-    this.refreshShouldEmit ||= emit
-    this.refreshing ??= this.runRefreshLoop().finally(() => {
-      this.refreshing = undefined
-      if (this.refreshPending && this.active) void this.refresh(false).catch(() => undefined)
-    })
-    return this.refreshing
-  }
-  private invalidate(): void {
-    if (this.refreshTimer || !this.active) return
-    this.refreshTimer = setTimeout(() => {
-      this.refreshTimer = undefined
-      void this.refresh(true).catch(() => undefined)
-    }, 16)
-  }
-  private requestAuthoritativeRefresh(): void {
-    if (this.refreshTimer) clearTimeout(this.refreshTimer)
-    this.refreshTimer = undefined
-    void this.refresh(true).catch(() => undefined)
   }
   private recordArtifacts(paths: readonly string[], agentId: string): void {
     const detail = this.detail
@@ -151,24 +126,19 @@ export class LoopalLiveSession {
     for (const artifact of created) this.sink.event({ type: 'artifact_created', artifact })
   }
 
-  private async runRefreshLoop(): Promise<void> {
-    while (this.refreshPending && this.active) {
-      this.refreshPending = false
-      const emit = this.refreshShouldEmit
-      this.refreshShouldEmit = false
-      const snapshot = await loadSessionDetail(
-        this.runtime.host, this.summaryValue, this.now, this.projector,
-        this.detailValue, this.attention.remoteAgentIds(),
-      )
-      if (!this.active) return
-      this.detailValue = { ...snapshot.detail, session: this.summaryValue }
-      this.attention.reconcile(
-        snapshot.pendingAttention,
-        new Set(snapshot.authoritativeRemoteAgents),
-      )
-      this.projector.finishSync(snapshot.revision, snapshot.revisions)
-      if (emit) this.sink.event({ type: 'session_detail_replaced', detail: this.detailValue })
-    }
+  private async refreshSnapshot(emit: boolean): Promise<void> {
+    const snapshot = await loadSessionDetail(
+      this.runtime.host, this.summaryValue, this.now, this.projector,
+      this.detailValue, this.attention.remoteAgentIds(),
+    )
+    if (!this.active) return
+    this.detailValue = { ...snapshot.detail, session: this.summaryValue }
+    this.attention.reconcile(
+      snapshot.pendingAttention,
+      new Set(snapshot.authoritativeRemoteAgents),
+    )
+    this.projector.finishSync(snapshot.revision, snapshot.revisions)
+    if (emit) this.sink.event({ type: 'session_detail_replaced', detail: this.detailValue })
   }
 
   private append(entry: ConversationEntry): void {
