+++ apps/desktop/src/platform/loopal-backend/node/sessions/loopal-session-snapshot.ts
@@ -1,11 +1,9 @@
 import {
   type AgentSummary,
-  type MetaHubRuntimeState,
   type SessionDetail,
   type SessionSummary,
 } from '../../../../shared/contracts'
 import { type DesktopHostClient } from '../backend/loopal-backend-types'
-import { type AttentionEventKind } from '../attention/loopal-attention'
 import { LoopalEventProjector } from '../projections/loopal-event-projector'
 import { projectMessages } from '../projections/loopal-message-projection'
 import {
@@ -17,11 +15,20 @@ import {
 import {
   TopologySchema,
   ViewSnapshotSchema,
-  AgentListSchema,
   type Topology,
   type ViewSnapshot,
 } from '../runtime/loopal-wire'
 import { mergeRemoteAgents, readMetaHubState } from '../federation/loopal-metahub-projection'
+import {
+  loadAgentSnapshots,
+  loadControllableAgents,
+  remoteAgentIds,
+  remoteSnapshotAuthority,
+  snapshotAttention,
+  type SnapshotAttention,
+} from './loopal-session-agent-snapshots'
+
+export { type SnapshotAttention } from './loopal-session-agent-snapshots'
 
 export async function loadSessionDetail(
   host: DesktopHostClient,
@@ -81,105 +88,6 @@ export async function loadSessionDetail(
       snapshotAttention(agentId, snapshot)
     )),
   }
-}
-
-async function loadControllableAgents(host: DesktopHostClient): Promise<ReadonlySet<string>> {
-  try {
-    const list = AgentListSchema.parse(await host.request('hub/list_agents', {}))
-    return new Set(list.agents.filter((agent) => (
-      agent.state === 'local' || agent.state === 'connected'
-    )).map((agent) => agent.name))
-  } catch {
-    return new Set()
-  }
-}
-
-export interface SnapshotAttention {
-  readonly kind: Extract<AttentionEventKind,
-    'permission_requested' | 'question_requested' | 'plan_approval_requested'>
-  readonly agentId: string
-  readonly value: unknown
-}
-
-function snapshotAttention(agentId: string, snapshot: ViewSnapshot): SnapshotAttention[] {
-  const conversation = snapshot.state.agent.conversation
-  const pending: SnapshotAttention[] = []
-  if (conversation.pending_permission) {
-    pending.push({
-      kind: 'permission_requested', agentId, value: conversation.pending_permission,
-    })
-  }
-  if (conversation.pending_question) {
-    const question = conversation.pending_question
-    pending.push({
-      kind: 'question_requested', agentId,
-      value: {
-        id: question.id,
-        questions: question.questions,
-        classifier_running: question.classifier_status?.kind === 'running',
-        classifier_status: question.classifier_status,
-      },
-    })
-  }
-  if (conversation.pending_plan_approval) {
-    pending.push({
-      kind: 'plan_approval_requested', agentId,
-      value: conversation.pending_plan_approval,
-    })
-  }
-  return pending
-}
-
-async function loadAgentSnapshots(
-  host: DesktopHostClient,
-  main: ViewSnapshot,
-  topology: Topology,
-  currentRemoteAgents: ReadonlySet<string>,
-  previousRemoteAgents: ReadonlySet<string>,
-): Promise<Map<string, ViewSnapshot>> {
-  const result = new Map<string, ViewSnapshot>([['main', main]])
-  const names = new Set([
-    ...topology.agents.map((agent) => agent.name).filter((name) => name !== 'main'),
-    ...currentRemoteAgents,
-    ...previousRemoteAgents,
-  ])
-  const snapshots = await Promise.all([...names].map(async (name) => {
-    try {
-      return [name, ViewSnapshotSchema.parse(
-        await host.request('view/snapshot', { agent: name }),
-      )] as const
-    } catch {
-      return undefined
-    }
-  }))
-  for (const snapshot of snapshots) {
-    if (snapshot) result.set(snapshot[0], snapshot[1])
-  }
-  return result
-}
-
-function remoteAgentIds(metaHub: MetaHubRuntimeState): Set<string> {
-  return new Set(metaHub.topology
-    .filter((agent) => agent.hub !== metaHub.hubName)
-    .map((agent) => agent.id))
-}
-
-function remoteSnapshotAuthority(
-  metaHub: MetaHubRuntimeState,
-  currentRemoteAgents: ReadonlySet<string>,
-  previousRemoteAgents: ReadonlySet<string>,
-  snapshots: ReadonlyMap<string, ViewSnapshot>,
-): string[] {
-  const authoritative = new Set([...snapshots.keys()].filter((agentId) => agentId.includes('/')))
-  if (metaHub.state !== 'connected') return [...authoritative]
-  const unavailableHubs = new Set(metaHub.topologyUnavailableHubs ?? [])
-  for (const agentId of previousRemoteAgents) {
-    const hub = agentId.split('/', 1)[0]
-    if (!currentRemoteAgents.has(agentId) && hub && !unavailableHubs.has(hub)) {
-      authoritative.add(agentId)
-    }
-  }
-  return [...authoritative]
 }
 
 function mergeAgents(
